+++ CHANGELOG.md
@@ -33,6 +33,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## Removed
  - Noob functions (and related variables and functions)
  - Superuser pull command
+ - Superuser restart command
 
 ## [1.2]
 ### Added